Small things: collision map, drag box,...
Drag box is no longer buggy, collision map is more efficient, unit
direction is now correct

+++ KingOfKings/src/Buildings/BuildingList.java
@@ -19,12 +19,12 @@ public class BuildingList implements IBuildingList {
 		return buildings.size();
 	}
 	
-	public int getBuildingX(int buildingNo){
+	public float getBuildingX(int buildingNo){
 		
 		return buildings.get(buildingNo).getX();
 	}
 	
-	public int getBuildingY(int buildingNo){
+	public float getBuildingY(int buildingNo){
 		
 		return buildings.get(buildingNo).getY();
 	}
